+++ include/fast_ber/ber_types/Construction.hpp
@@ -18,7 +18,7 @@ constexpr inline Construction get_construction(uint8_t first_byte)
 
 constexpr inline uint8_t add_construction(uint8_t first_byte, Construction construction)
 {
-    return (static_cast<int>(construction) << 5) | (first_byte & 0xDF);
+    return (static_cast<uint8_t>(construction) << 5) | (first_byte & 0xDF);
 }
 
 } // namespace fast_ber
